add security for modify only if logedin as admin

+++ pages/api/products/[id].jsx
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+
+  const token = cookies.token;
 
   await dbConnect();
 
@@ -19,6 +22,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authentificated");
+    }
     try {
       const updatedProduct = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -30,6 +36,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authentificated");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json({ message: "Product deleted successfully!" });
